Enable warning 4389 as an error on windows builds. This will make windows builds more similar to linux/mac, which already treat signed/ unsigned equality comparisons as warnings (and hence errors).
BUG=44471
TEST=none 

Review URL: http://codereview.chromium.org/2222002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@48395 0039d316-1c4b-4281-b951-d872f2087c98

+++ base/thread_local.h
@@ -1,5 +1,5 @@
 FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
-// Copyright (c) 2006-2008 The Chromium Authors. All rights reserved.
+// Copyright (c) 2010 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -60,7 +60,7 @@ namespace base {
 // Helper functions that abstract the cross-platform APIs.  Do not use directly.
 struct ThreadLocalPlatform {
 #if defined(OS_WIN)
-  typedef int SlotType;
+  typedef unsigned long SlotType;
 #elif defined(OS_POSIX)
   typedef pthread_key_t SlotType;
 #endif
